LayoutTests/platform/glib/imported/w3c/web-platform-tests/editing/other/insertparagraph-in-child-of-head.tentative_designMode=on&white-space=normal-expected.txt



PASS insertParagraph in empty <title style="display:block;white-space:normal"> should not split it
PASS insertParagraph in <title style="display:block;white-space:normal"> containing text should not split it
PASS insertParagraph in empty <style style="display:block;white-space:normal"> should not split it
PASS insertParagraph in <style style="display:block;white-space:normal"> containing text should not split it
PASS insertParagraph in empty <script style="display:block;white-space:normal"> should not split it
PASS insertParagraph in <script style="display:block;white-space:normal"> containing text should not split it
FAIL insertParagraph in empty <div style="white-space:normal"> in the <head> should split the <div> assert_equals: The <div> should be split expected "<head><title>iframe</title><div style=\"white-space:normal\"><br></div><div style=\"white-space:normal\"><br></div></head><body><br></body>" but got "<head><title>iframe</title><div style=\"white-space:normal\"></div></head><body><br></body>"
FAIL insertParagraph in <div style="white-space:normal"> (containing only a <br>) in the <head> should split the <div> element assert_equals: The <div> should be split expected "<head><title>iframe</title><div style=\"white-space:normal\"><br></div><div style=\"white-space:normal\"><br></div></head><body><br></body>" but got "<head><title>iframe</title><div style=\"white-space:normal\"><br></div></head><body><br></body>"
FAIL insertParagraph in <div style="white-space:normal"> (containing text) in the <head> should split the <div> element assert_in_array: The <div> should be split value "<head><title>iframe</title><div style=\"white-space:normal\">abcd</div></head><body><br></body>" not in array ["<head><title>iframe</title><div style=\"white-space:normal\">ab</div><div style=\"white-space:normal\">cd</div></head><body><br></body>", "<head><title>iframe</title><div style=\"white-space:normal\">ab</div><div style=\"white-space:normal\">cd<br></div></head><body><br></body>", "<head><title>iframe</title><div style=\"white-space:normal\">ab<br></div><div style=\"white-space:normal\">cd</div></head><body><br></body>", "<head><title>iframe</title><div style=\"white-space:normal\">ab<br></div><div style=\"white-space:normal\">cd<br></div></head><body><br></body>"]
